Refactor MeterName from class to interface
